ai-backend/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
